remove more parts of tcpdi
