+++ lib/Command/Configure/Cfssl.php
@@ -5,14 +5,10 @@ declare(strict_types=1);
 namespace OCA\Libresign\Command\Configure;
 
 use InvalidArgumentException;
-use OC\SystemConfig;
 use OCA\Libresign\AppInfo\Application;
 use OCA\Libresign\Command\Base;
 use OCA\Libresign\Service\AdminSignatureService;
-use OCP\Files\IRootFolder;
-use OCP\Http\Client\IClientService;
-use OCP\IConfig;
-use OCP\ITempManager;
+use OCA\Libresign\Service\InstallService;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -21,19 +17,11 @@ class Cfssl extends Base {
 	/** @var AdminSignatureService */
 	private $adminSignatureService;
 	public function __construct(
-		ITempManager $tempManager,
-		IClientService $clientService,
-		IConfig $config,
-		SystemConfig $systemConfig,
-		IRootFolder $rootFolder,
+		InstallService $installService,
 		AdminSignatureService $adminSignatureService
 	) {
 		parent::__construct(
-			$tempManager,
-			$clientService,
-			$config,
-			$systemConfig,
-			$rootFolder
+			$installService
 		);
 		$this->adminSignatureService = $adminSignatureService;
 	}
@@ -93,7 +81,7 @@ class Cfssl extends Base {
 		if (!$country = $input->getOption('c')) {
 			throw new InvalidArgumentException('Invalid Country');
 		}
-		if ($binary = $this->config->getAppValue(Application::APP_ID, 'cfssl_bin')) {
+		if ($binary = $this->installService->config->getAppValue(Application::APP_ID, 'cfssl_bin')) {
 			if (PHP_OS_FAMILY === 'Windows') {
 				throw new InvalidArgumentException('Incompatible with Windows');
 			}
@@ -104,8 +92,8 @@ class Cfssl extends Base {
 				throw new InvalidArgumentException('CFSSL URI is not necessary');
 			}
 			// create if not exist
-			$this->getFolder('cfssl_config');
-			$configPath = $this->getFullPath() . DIRECTORY_SEPARATOR . 'cfssl_config' . DIRECTORY_SEPARATOR;
+			$this->installService->getFolder('cfssl_config');
+			$configPath = $this->installService->getFullPath() . DIRECTORY_SEPARATOR . 'cfssl_config' . DIRECTORY_SEPARATOR;
 			$cfsslUri = 'http://127.0.0.1:8888/api/v1/cfssl/';
 		} else {
 			$output->writeln('CFSSL binary not found!');
