+++ lib/Command/Install.php
@@ -4,29 +4,17 @@ declare(strict_types=1);
 
 namespace OCA\Libresign\Command;
 
-use OC\SystemConfig;
-use OCP\Files\IRootFolder;
-use OCP\Http\Client\IClientService;
-use OCP\IConfig;
-use OCP\ITempManager;
+use OCA\Libresign\Service\InstallService;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class Install extends Base {
 	public function __construct(
-		ITempManager $tempManager,
-		IClientService $clientService,
-		IConfig $config,
-		SystemConfig $systemConfig,
-		IRootFolder $rootFolder
+		InstallService $installService
 	) {
 		parent::__construct(
-			$tempManager,
-			$clientService,
-			$config,
-			$systemConfig,
-			$rootFolder
+			$installService
 		);
 	}
 
@@ -64,20 +52,21 @@ class Install extends Base {
 	protected function execute(InputInterface $input, OutputInterface $output): int {
 		$all = $input->getOption('all');
 		$ok = false;
+		$this->installService->setOutput($output);
 		if ($input->getOption('java') || $all) {
-			$this->installJava($output);
+			$this->installService->installJava();
 			$ok = true;
 		}
 		if ($input->getOption('jsignpdf') || $all) {
-			$this->installJSignPdf($output);
+			$this->installService->installJSignPdf();
 			$ok = true;
 		}
 		if ($input->getOption('cfssl') || $all) {
-			$this->installCfssl($output);
+			$this->installService->installCfssl();
 			$ok = true;
 		}
 		if ($input->getOption('cli') || $all) {
-			$this->installCli($output);
+			$this->installService->installCli();
 			$ok = true;
 		}
 		if (!$ok) {
